requirements.txt
asgiref==3.10.0
Django==5.2.7
django-extensions==4.1
djangorestframework==3.16.1
drf-yasg==1.21.11
inflection==0.5.1
packaging==25.0
psycopg2-binary==2.9.11
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.3
uritemplate==4.2.0
